feat(Card)!: remove top-level sub-component(s) (#1692)
- Combine card sub components into one file
- update docs to reflect this new paradigm
- rename directory and adjust copy
