--- src/components/CardBody/index.ts
@@ -1,1 +1,0 @@
-export { CardBody as default } from './CardBody';
