--- src/components/CardFooter/index.ts
@@ -1,1 +1,0 @@
-export { CardFooter as default } from './CardFooter';
