--- src/components/CardHeader/index.ts
@@ -1,1 +1,0 @@
-export { CardHeader as default } from './CardHeader';
