+++ src/index.ts
@@ -10,9 +10,6 @@ export { default as Button } from './components/Button';
 export { default as ButtonDropdown } from './components/ButtonDropdown';
 export { default as ButtonGroup } from './components/ButtonGroup';
 export { default as Card } from './components/Card';
-export { default as CardBody } from './components/CardBody';
-export { default as CardFooter } from './components/CardFooter';
-export { default as CardHeader } from './components/CardHeader';
 export { default as Checkbox } from './components/Checkbox';
 export { default as ClickableStyle } from './components/ClickableStyle';
 export { default as DataBar } from './components/DataBar';
